# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "TechFuturist API Documentation",
        "description": "![TechFuturist Logo](http://localhost:8000/images/images.png)\n * Documentation de l'API pour l'application TechFuturist",
        "contact": {
            "email": "support@techfuturist.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Serveur principal"
        }
    ],
    "paths": {
...
        "/api/tests": {
            "get": {
                "tags": [
                    "Tests"
                ],
                "summary": "Liste de tous les tests",
                "description": "Retourne la liste complète des tests. Accessible à tous les utilisateurs authentifiés.",
                "operationId": "c111461571e8c1952d5ec29f6b79e9ce",
                "responses": {
                    "200": {
                        "description": "Liste des tests",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Test"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Tests"
                ],
                "summary": "Créer un nouveau test",
                "description": "Seuls les administrateurs peuvent créer un test.",
                "operationId": "ce2fcda95b04246f42b319d2f2dce318",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "title",
                                    "type",
                                    "file",
                                    "file_type"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Test QCM PHP"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Description détaillée"
                                    },
                                    "type": {
                                        "type": "string",
                                        "enum": [
                                            "QCM",
                                            "Ouvert",
                                            "Pratique"
                                        ],
                                        "example": "QCM"
                                    },
                                    "file": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "file_type": {
                                        "type": "string",
                                        "example": "pdf"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Test créé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Test"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Accès refusé"
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/tests/{id}": {
            "get": {
                "tags": [
                    "Tests"
                ],
                "summary": "Afficher un test spécifique",
                "description": "Retourne les détails d'un test donné.",
                "operationId": "280700573d2da846da72854c8b95aabc",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du test",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détails du test",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Test"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Test non trouvé"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Tests"
                ],
                "summary": "Mettre à jour un test",
                "description": "Seuls les administrateurs peuvent modifier un test.",
                "operationId": "7fd12f63a4a495d43fde843fe17f3d30",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du test à modifier",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Test QCM modifié"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Nouvelle description"
                                    },
                                    "type": {
                                        "type": "string",
                                        "enum": [
                                            "QCM",
                                            "Ouvert",
                                            "Pratique"
                                        ]
                                    },
                                    "file": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "file_type": {
                                        "type": "string",
                                        "example": "pdf"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Test mis à jour",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Test"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Accès refusé"
                    },
                    "404": {
                        "description": "Test non trouvé"
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Tests"
                ],
                "summary": "Supprimer un test",
                "description": "Seuls les administrateurs peuvent supprimer un test.",
                "operationId": "e05e41a3deb85323a1267599bbc9c472",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du test à supprimer",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Test supprimé"
                    },
                    "403": {
                        "description": "Accès refusé"
                    },
                    "404": {
                        "description": "Test non trouvé"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/tests/download/{id}": {
            "get": {
                "tags": [
                    "Tests"
                ],
                "summary": "Télécharger un test",
                "description": "Tous les utilisateurs peuvent télécharger un test.",
                "operationId": "9a0b1026c8608742d8bb1777edae068c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du test à télécharger",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Fichier téléchargé"
                    },
                    "403": {
                        "description": "Accès refusé"
                    },
                    "404": {
                        "description": "Test ou fichier non trouvé"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Test": {
                "title": "Test",
                "required": [
                    "id",
                    "title",
                    "type",
                    "file_url",
                    "file_type",
                    "created_by",
                    "created_at",
                    "updated_at"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "example": "Test QCM sur PHP"
                    },
                    "description": {
                        "type": "string",
                        "example": "Description détaillée du test",
                        "nullable": true
                    },
                    "type": {
                        "type": "string",
                        "enum": [
                            "QCM",
                            "Ouvert",
                            "Pratique"
                        ],
                        "example": "QCM"
                    },
                    "file_url": {
                        "type": "string",
                        "example": "uploads/tests/abc123.pdf"
                    },
                    "file_type": {
                        "type": "string",
                        "example": "pdf"
                    },
                    "created_by": {
                        "type": "integer",
                        "example": 42
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-08-10T12:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-08-11T12:00:00Z"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}